Calculate merged values for regs, cmem, lmem and smem during analysis
Analysis is done in a per-kernel basis, meaning that Register, Local, Shared
and Constant memory usage is calculated for each kernel. This was
left unimplemented for the L1I cache injections, as the refactored
scripts had no need for them. For completeness, the "total" or
"merged" values are now calculated too, meaning that the overall
value of registers, cmem, lmem and smem usage over all kernels is
now stored into the merged_executable_analysis.sh file.

For all values, the maximum value over all kernels is kept.